Use composable details for proposal preview
