+++ packages/ui/src/common/utils/downloadFile.tsx
@@ -1,0 +1,6 @@
+export function downloadFile(filename: string, downloadHref: string) {
+  const a = document.createElement('a')
+  a.download = filename
+  a.href = downloadHref
+  a.click()
+}
