+++ packages/ui/src/proposals/components/ProposalDetails/ProposalDetails.stories.tsx
@@ -4,39 +4,42 @@ import React from 'react'
 import { MemoryRouter } from 'react-router-dom'
 
 import { RowGapBlock } from '@/common/components/page/PageContent'
+import { MockApolloProvider } from '@/mocks/components/storybook/MockApolloProvider'
 
 import { getMember } from '../../../../test/_mocks/members'
 import { ModalContext } from '../../../common/providers/modal/context'
 
-import { ProposalDetailsComponent } from './ProposalDetails'
+import { ProposalDetails } from './ProposalDetails'
 
 export default {
   title: 'Proposals/ProposalPreview/ProposalDetails',
-  component: ProposalDetailsComponent,
+  component: ProposalDetails,
 } as Meta
 
-type Props = Parameters<typeof ProposalDetailsComponent>[0]
+type Props = Parameters<typeof ProposalDetails>[number]
 
 const Template: Story<Props> = (args) => (
   <MemoryRouter>
-    <ModalContext.Provider
-      value={{
-        modal: null,
-        modalData: null,
-        showModal: () => undefined,
-        hideModal: () => undefined,
-      }}
-    >
-      <RowGapBlock gap={24}>
-        <ProposalDetailsComponent {...args} />
-      </RowGapBlock>
-    </ModalContext.Provider>
+    <MockApolloProvider members proposals workingGroups workers>
+      <ModalContext.Provider
+        value={{
+          modal: null,
+          modalData: null,
+          showModal: () => undefined,
+          hideModal: () => undefined,
+        }}
+      >
+        <RowGapBlock gap={24}>
+          <ProposalDetails {...args} />
+        </RowGapBlock>
+      </ModalContext.Provider>
+    </MockApolloProvider>
   </MemoryRouter>
 )
 
 export const CreateLeadOpening = Template.bind({})
 CreateLeadOpening.args = {
-  details: {
+  proposalDetails: {
     type: 'createWorkingGroupLeadOpening',
     stakeAmount: new BN(10000),
     unstakingPeriod: new BN(14400),
@@ -50,14 +53,26 @@ CreateLeadOpening.args = {
   },
 }
 
+export const RuntimeUpgrade = Template.bind({})
+RuntimeUpgrade.args = {
+  proposalDetails: {
+    type: 'runtimeUpgrade',
+    newBytecodeId: '1',
+  },
+}
+
 export const FundingRequest = Template.bind({})
 FundingRequest.args = {
-  details: {
+  proposalDetails: {
     type: 'fundingRequest',
     destinations: [
       {
         account: '5GBVQJZzv2nrRVtjgTg4JwCwbEfKeueiKqQ5ysFxLfiQPRiY',
         amount: 30000,
+      },
+      {
+        account: '5JwCwbEfKeueiKqQ5ysFxLfiQPRiYGBVQJZzv2nrRVtjgTg4',
+        amount: 140000,
       },
       {
         account: '5JwCwbEfKeueiKqQ5ysFxLfiQPRiYGBVQJZzv2nrRVtjgTg4',
@@ -69,7 +84,7 @@ FundingRequest.args = {
 
 export const DecreaseLeadStake = Template.bind({})
 DecreaseLeadStake.args = {
-  details: {
+  proposalDetails: {
     type: 'decreaseWorkingGroupLeadStake',
     groupName: 'storage',
     amount: new BN(10000),
@@ -79,7 +94,7 @@ DecreaseLeadStake.args = {
 
 export const SlashLead = Template.bind({})
 SlashLead.args = {
-  details: {
+  proposalDetails: {
     type: 'slashWorkingGroupLead',
     groupName: 'forum',
     amount: new BN(15000),
