+++ packages/ui/src/proposals/components/ProposalDetails/renderers/Address.tsx
@@ -1,9 +1,7 @@
 import React from 'react'
 
-import { CopyComponent } from '@/common/components/CopyComponent'
-import { Row } from '@/common/components/Modal'
-import { RowGapBlock } from '@/common/components/page/PageContent'
-import { Label } from '@/common/components/typography'
+import { CopyButton } from '@/common/components/buttons'
+import { NumericValue, StatiscticSpaceRow, StatisticItem } from '@/common/components/statistics'
 import { shortenAddress } from '@/common/model/formatters'
 
 interface Props {
@@ -12,10 +10,10 @@ interface Props {
 }
 
 export const Address = ({ label, value }: Props) => (
-  <Row>
-    <RowGapBlock gap={4}>
-      <Label>{label}</Label>
-      <CopyComponent altText={shortenAddress(value)} copyText={value} />
-    </RowGapBlock>
-  </Row>
+  <StatisticItem title={label}>
+    <StatiscticSpaceRow>
+      <NumericValue>{shortenAddress(value, 12)}</NumericValue>
+      <CopyButton textToCopy={value} />
+    </StatiscticSpaceRow>
+  </StatisticItem>
 )
