+++ packages/ui/src/proposals/components/ProposalDetails/renderers/Divider.tsx
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const Divider = () => <div />
