+++ packages/ui/src/proposals/components/ProposalDetails/renderers/Markdown/components/index.ts
@@ -1,0 +1,1 @@
+export * from './OpeningDescriptionPreview'
