+++ packages/ui/src/proposals/components/ProposalDetails/renderers/NumberOfBlocks.tsx
@@ -1,9 +1,8 @@
 import BN from 'bn.js'
 import React from 'react'
 
-import { Row } from '@/common/components/Modal'
-import { RowGapBlock } from '@/common/components/page/PageContent'
-import { Label, TextInlineMedium, TextMedium } from '@/common/components/typography'
+import { StatisticItem } from '@/common/components/statistics'
+import { TextInlineBig } from '@/common/components/typography'
 import { formatTokenValue } from '@/common/model/formatters'
 
 interface Props {
@@ -12,12 +11,9 @@ interface Props {
 }
 
 export const NumberOfBlocks = ({ label, value }: Props) => (
-  <Row>
-    <RowGapBlock gap={4}>
-      <Label>{label}</Label>
-      <TextMedium lighter>
-        <TextInlineMedium dark>{formatTokenValue(value)}</TextInlineMedium> blocks
-      </TextMedium>
-    </RowGapBlock>
-  </Row>
+  <StatisticItem title={label}>
+    <TextInlineBig bold value>
+      {formatTokenValue(value)} blocks
+    </TextInlineBig>
+  </StatisticItem>
 )
