+++ packages/ui/src/proposals/components/ProposalDetails/renderers/index.ts
@@ -1,7 +1,8 @@
 export * from './Text'
 export * from './Amount'
 export * from './NumberOfBlocks'
-export * from './Markdown'
+export * from './Markdown/Markdown'
 export * from './Member'
 export * from './Address'
 export * from './RuntimeBlob'
+export * from './Divider'
